Just add location data to rsx items
