+++ packages/rsx/src/location.rs
@@ -1,0 +1,8 @@
+/// Information about the location of the call to a component
+///
+/// This will be filled in when the dynamiccontext is built, filling in the file:line:column:id format
+///
+#[derive(PartialEq, Eq, Clone, Debug, Hash, Default)]
+pub struct CallerLocation {
+    inner: Option<String>,
+}
